Resolvi os erros

+++ auxiliares/dados.py
@@ -74,6 +74,6 @@ class Dados:
     novo_df = arquivo_csv
     novo_df.set_index('código', drop=True, inplace=True)
     novo_df = novo_df.drop(columns=["Frete", "unidade", "Rendimento %", "Custo und."])
-    novo_df["Preço"] = novo_df["Preço"].replace(',','.', regex=True)
+    novo_df["Preço"] = novo_df["Preço"].replace(',','.', regex=True).astype(float)
 
     return novo_df
